Screen refresh works now

+++ src/gameEngine/gameEngine.h
@@ -1,13 +1,14 @@
+#include <stdint.h>
 #ifndef GAMEENGINE_H
 #define GAMEENGINE_H
 
 #define GRID_SIZE 4
 typedef struct gameState {
-    int** grid;			 // 2D grid (allocate carefully)
-    int currentMove;	 // e.g., 0=up, 1=right, 2=down, 3=left
-    int moveSuccessful;	 // 1 if move changed the grid, else 0
-    int gameIsOver;
-    int biggestSquare;
+    uint8_t** grid;			 // 2D grid (allocate carefully)
+    uint8_t currentMove;	 // e.g., 0=up, 1=right, 2=down, 3=left
+    uint8_t moveSuccessful;	 // 1 if move changed the grid, else 0
+    uint8_t gameIsOver;
+    uint8_t biggestSquare;
 } gameState;
 
 struct gameState* newGame();
